Auth añadido, Fix Entity User, SQL Cambiada, DAO User y AuthToken

+++ football-service/src/main/java/edu/upc/eetac/dsa/football/entity/User.java
@@ -52,7 +52,7 @@ public class User {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(float balance) {
         this.balance = balance;
     }
 }
